refactor: move codeframe to @vue/shared
