+++ packages/shared/src/index.ts
@@ -3,6 +3,7 @@ import { makeMap } from './makeMap'
 export { makeMap }
 export * from './patchFlags'
 export { isGloballyWhitelisted } from './globalsWhitelist'
+export { generateCodeFrame } from './codeframe'
 
 export const EMPTY_OBJ: { readonly [key: string]: any } = __DEV__
   ? Object.freeze({})
